Move SingleCertAndKey to crypto::signer
